sim.py: OPT uses LUT to run in ~ O(|S|^K K N) when K small
- Pre-calculation of previous arrivals effectively runs the alg
in an iterative/incremental fashion, and avoids deep nested recursion.
- Note that when K is small compared to |S|, the offline_opt_recursion
routine is more likely to find the current request r is NOT in the set
of edge services, thus avoiding the min operation of |S|-K+1 choices.
- Since it is incremental, LUT in fact only needs to hold the result
with the last/previous arrival. In this way the space complexity can
be reduced, if necessary.